# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "E-commerce API",
    "description": "Documentation for e-commerce APIMy API"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/memberships/": {
      "get": {
        "tags": [
          "Memberships"
        ],
        "description": "Get All Memberships",
        "produces": [
          "text/json"
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "result": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "number",
                                "example": 1
                              },
                              "name": {
                                "type": "string",
                                "example": "Bronze"
                              },
                              "createdAt": {
                                "type": "string",
                                "example": "2023-11-28T15:59:06.000Z"
                              },
                              "updatedAt": {
                                "type": "string",
                                "example": "2023-11-28T15:59:06.000Z"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Memberships"
        ],
        "description": "Add a Membership",
        "produces": [
          "text/json"
        ],
        "parameters": [
          {
            "name": "user-agent",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Bronze"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 201
                        },
                        "result": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "number",
                              "example": 7
                            },
                            "name": {
                              "type": "string",
                              "example": "Bronze"
                            },
                            "updatedAt": {
                              "type": "string",
                              "example": "2023-11-30T07:37:23.428Z"
                            },
                            "createdAt": {
                              "type": "string",
                              "example": "2023-11-30T07:37:23.428Z"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Created"
          },
          "400": {
            "description": "All fields are required"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/memberships/{membershipId}": {
      "delete": {
        "tags": [
          "Memberships"
        ],
        "description": "Delete a Membership",
        "produces": [
          "text/json"
        ],
        "parameters": [
          {
            "name": "membershipId",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Membership Id",
            "example": 1
          },
          {
            "name": "user-agent",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Membership deleted"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "The Membership is being used"
          }
        }
      },
      "put": {
        "tags": [
          "Memberships"
        ],
        "description": "Edit a Membership",
        "produces": [
          "text/json"
        ],
        "parameters": [
          {
            "name": "membershipId",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Membership Id",
            "example": 1
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "test1"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "result": {
                          "type": "string",
                          "example": "Membership updated"
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "Name is required"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}